+++ Server/StreetEmpire.Api/Migrations/20260817013808_Lookout.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace StreetEmpire.Api.Migrations
+{
+    /// <inheritdoc />
+    public partial class Lookout : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "LookoutLevel",
+                table: "Hideouts",
+                type: "integer",
+                nullable: false,
+                defaultValue: 0);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "LookoutLevel",
+                table: "Hideouts");
+        }
+    }
+}